fixed constuctor

+++ CheckingAccount.java
@@ -3,7 +3,7 @@ public class CheckingAccount extends Account {
     private static final double MIN_BALANCE = 50;
     private static final double INTEREST_RATE = 0;
 
-    public SavingAccount(int account_id, int pin, int balance) {
+    public SavingAccount(int account_id, int pin, double balance) {
 	super(account_id, pin, balance);
     }
