+++ SavingAccount.java
@@ -3,7 +3,7 @@ public class SavingAccount extends Account {
     private static final double MIN_BALANCE = 200;
     private static final double INTEREST_RATE = 0.005 / 12;
 
-    public SavingAccount(int account_id, int pin, int balance) {
+    public SavingAccount(int account_id, int pin, double balance) {
 	super(account_id, pin, balance);
     }
